refactor(open-api): parser owns the fully-linked model graph
buildClientData now returns a fully-linked, composite-resolved graph:
model linking and composite resolution run inside the parser, and
codegen-data.ts is purely the augmentation + language-rendering stage.

- Delete buildModelForPrimitive's throwaway-spec hack (a synthetic spec +
  sentinel schema name re-run per nested value); buildOrReferenceModel
  reuses the parser's buildInlineModel + linkModel directly.
- Set type=name for object definitions at construction in the parser.
- Move linkModel/linkModels, resolveComposedModels and the shared spec
  helpers into parser.ts.
- Add unit coverage for the inline value-model path (additionalProperties,
  patternProperties, jsonl itemSchema) and composite member contents.

Generated client, hooks and metadata remain byte-identical (233 open-api
tests, zero snapshot changes).
